requirements.txt
streamlit==1.50.0
python-dotenv==1.1.1
langchain==0.3.27
langchain-community==0.3.29
langchain-core==0.3.76
langchain-groq==0.3.8
langchain-huggingface==0.3.1
faiss-cpu==1.12.0
tiktoken==0.11.0
transformers==4.56.2
sentence-transformers==5.1.1
PyPDF2==3.0.1






